apps/kpi-functional-dashboard-backend/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64"]

[project]
name = "nat-react-agent-blueprint"
dynamic = ["version"]
dependencies = [
  "nvidia-nat[telemetry, langchain, mcp]==1.3.0",
  "pydantic>=2.10.0",
  "langchain-core>=0.3.0",
  "langchain>=0.3.0",
  "requests>=2.32.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=1.0.0",
  "PyJWT>=2.8.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "SQLAlchemy>=2.0.0",
  "PyYAML>=6.0.0",
  "psycopg2-binary>=2.9.0",  # PostgreSQL database adapter for Vanna
  "pymilvus>=2.4.0",  # Milvus vector database client for production deployments
  
  # Libraries that require exact versions for NVIDIA Embedding function fixes
  "langchain-nvidia-ai-endpoints==0.3.18",
  "chromadb==1.1.0",
  "vanna==0.7.9",
]
requires-python = ">=3.11"
description = "NAT React Agent Blueprint for Astra"
classifiers = ["Programming Language :: Python"]
authors = [{ name = "NVIDIA Corporation" }]
maintainers = [{ name = "NVIDIA Corporation" }]

[project.entry-points.'nat.components']
nat_react_agent_blueprint = "nat_react_agent_blueprint.register"

[tool.setuptools]
packages = ["nat_react_agent_blueprint"]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "nat_react_agent_blueprint.__version__"}

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "e2e: end-to-end tests that run full workflows",
    "integration: integration tests with external APIs",
    "unit: unit tests for individual components",
]
testpaths = [
    ".",
]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
